+++ packages/bloc/example/main.dart
@@ -16,6 +16,12 @@ class SimpleBlocObserver extends BlocObserver {
   }
 
   @override
+  void onChange(BlocBase bloc, Change change) {
+    super.onChange(bloc, change);
+    print('onChange -- bloc: ${bloc.runtimeType}, change: $change');
+  }
+
+  @override
   void onTransition(BlocBase bloc, Transition transition) {
     super.onTransition(bloc, transition);
     print('onTransition -- bloc: ${bloc.runtimeType}, transition: $transition');