+++ packages/bloc/lib/src/transition.dart
@@ -1,26 +1,52 @@
 import 'package:meta/meta.dart';
 
+/// {@template change}
+/// A [Change] represents the change from one [State] to another.
+/// A [Change] consists of the [currentState] and [nextState].
+/// {@endtemplate}
+@immutable
+class Change<State> {
+  /// {@macro change}
+  const Change({required this.currentState, required this.nextState});
+
+  /// The current [State] at the time of the [Change].
+  final State currentState;
+
+  /// The next [State] at the time of the [Change].
+  final State nextState;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Change<State> &&
+          runtimeType == other.runtimeType &&
+          currentState == other.currentState &&
+          nextState == other.nextState;
+
+  @override
+  int get hashCode => currentState.hashCode ^ nextState.hashCode;
+
+  @override
+  String toString() {
+    return 'Change { currentState: $currentState, nextState: $nextState }';
+  }
+}
+
 /// {@template transition}
 /// A [Transition] is the change from one state to another.
 /// Consists of the [currentState], an [event], and the [nextState].
 /// {@endtemplate}
 @immutable
-class Transition<Event, State> {
+class Transition<Event, State> extends Change<State> {
   /// {@macro transition}
   const Transition({
-    required this.currentState,
+    required State currentState,
     required this.event,
-    required this.nextState,
-  });
-
-  /// The current [State] at the time of the [Transition].
-  final State currentState;
+    required State nextState,
+  }) : super(currentState: currentState, nextState: nextState);
 
   /// The [Event] which triggered the current [Transition].
   final Event event;
-
-  /// The next [State] at the time of the [Transition].
-  final State nextState;
 
   @override
   bool operator ==(Object other) =>
@@ -38,6 +64,6 @@ class Transition<Event, State> {
 
   @override
   String toString() {
-    return '''Transition { currentState: $currentState, ${event == null ? '' : 'event: $event, '}nextState: $nextState }''';
+    return '''Transition { currentState: $currentState, event: $event, nextState: $nextState }''';
   }
 }