+++ packages/bloc_test/lib/src/mock_bloc.dart
@@ -23,7 +23,7 @@ import 'package:mocktail/mocktail.dart';
 /// class MockCounterBloc extends MockBloc implements CounterBloc {}
 /// ```
 /// {@endtemplate}
-class MockBloc<E, S> extends _MockBlocBase<E, S> implements Bloc<E, S> {
+class MockBloc<E, S> extends _MockBlocBase<S> implements Bloc<E, S> {
   /// {@macro mock_bloc}
   MockBloc() {
     when(() => mapEventToState(any())).thenAnswer((_) => Stream<S>.empty());
@@ -53,9 +53,9 @@ class MockBloc<E, S> extends _MockBlocBase<E, S> implements Bloc<E, S> {
 /// class MockCounterCubit extends MockBloc implements CounterCubit {}
 /// ```
 /// {@endtemplate}
-class MockCubit<S> extends _MockBlocBase<Null, S> implements Cubit<S> {}
+class MockCubit<S> extends _MockBlocBase<S> implements Cubit<S> {}
 
-class _MockBlocBase<E, S> extends Mock implements BlocBase<E, S> {
+class _MockBlocBase<S> extends Mock implements BlocBase<S> {
   _MockBlocBase() {
     registerFallbackValue<void Function(S)>((S _) {});
     registerFallbackValue<void Function()>(() {});